+++ execution/market_snapshot.go
@@ -70,6 +70,12 @@ func NewMarketFromSnapshot(
 		mkt.ID,
 		asset,
 		stateVarEngine,
+		&types.RiskFactor{
+			Market: mkt.ID,
+			Short:  em.ShortRiskFactor,
+			Long:   em.LongRiskFactor,
+		},
+		em.RiskFactorConsensusReached,
 	)
 
 	settleEngine := settlement.New(
@@ -95,7 +101,7 @@ func NewMarketFromSnapshot(
 
 	lMonitor := lmon.NewMonitor(tsCalc, mkt.LiquidityMonitoringParameters)
 
-	liqEngine := liquidity.NewSnapshotEngine(liquidityConfig, log, broker, idgen, tradableInstrument.RiskModel, pMonitor, mkt.ID)
+	liqEngine := liquidity.NewSnapshotEngine(liquidityConfig, log, broker, idgen, tradableInstrument.RiskModel, pMonitor, asset, mkt.ID, stateVarEngine, mkt.TickSize())
 	// call on chain time update straight away, so
 	// the time in the engine is being updatedat creation
 	liqEngine.OnChainTimeUpdate(ctx, now)
@@ -131,7 +137,7 @@ func NewMarketFromSnapshot(
 		markPrice:          em.CurrentMarkPrice.Clone(),
 		stateChanged:       true,
 	}
-
+	liqEngine.SetGetStaticPricesFunc(market.getBestStaticPrices)
 	return market, nil
 }
 
@@ -148,6 +154,7 @@ func (m *Market) changed() bool {
 }
 
 func (m *Market) getState() *types.ExecMarket {
+	rf, _ := m.risk.GetRiskFactors()
 	em := &types.ExecMarket{
 		Market:                     m.mkt.DeepClone(),
 		PriceMonitor:               m.pMonitor.GetState(),
@@ -162,6 +169,9 @@ func (m *Market) getState() *types.ExecMarket {
 		CurrentMarkPrice:           m.getCurrentMarkPrice(),
 		LastEquityShareDistributed: m.lastEquityShareDistributed.Unix(),
 		EquityShare:                m.equityShares.GetState(),
+		RiskFactorConsensusReached: m.risk.IsRiskFactorInitialised(),
+		ShortRiskFactor:            rf.Short,
+		LongRiskFactor:             rf.Long,
 	}
 
 	m.stateChanged = false